# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Verde2 transactional email",
    "version": "2.0.0",
    "description": "One tenant per application/environment. Bearer credentials are opaque and expiring. API default scope send:template. SMTP credentials separate. No attachments; one recipient. Templates support escaped declared simple variables only. Internal endpoints are only available on the private TLS listener. Idempotency lasts 30 days."
  },
  "paths": {
    "/api/v1/send": {
      "post": {
        "operationId": "post_api_v1_send",
        "description": "Persisted and accepted for processing, not delivered. Idempotency window 30 days. While dispatch is paused, new admissions return 503 dispatch_paused without persistence. An existing matching idempotency key returns its message; a conflicting payload returns 409.",
        "tags": [
          "application"
        ],
        "security": [
          {
            "Application": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "to": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 254,
                    "format": "email"
                  },
                  "templateId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "variables": {
                    "type": "object",
                    "maxProperties": 50,
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "string",
                          "maxLength": 10000
                        },
                        {
                          "type": "number"
                        }
                      ]
                    },
                    "propertyNames": {
                      "pattern": "^[A-Za-z][A-Za-z0-9_]{0,63}$"
                    }
                  }
                },
                "required": [
                  "to",
                  "templateId"
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Persisted and accepted for processing. Not delivery confirmation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "state": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 30
                    }
                  },
                  "required": [
                    "id",
                    "state"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/send/raw": {
      "post": {
        "operationId": "post_api_v1_send_raw",
        "description": "Persisted and accepted for processing, not delivered. Idempotency window 30 days. While dispatch is paused, new admissions return 503 dispatch_paused without persistence. An existing matching idempotency key returns its message; a conflicting payload returns 409.",
        "tags": [
          "application"
        ],
        "security": [
          {
            "Application": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "from": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 254,
                    "format": "email"
                  },
                  "replyTo": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 254,
                    "format": "email"
                  },
                  "subject": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 998
                  },
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000000
                  },
                  "html": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000000
                  },
                  "to": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 254,
                    "format": "email"
                  }
                },
                "required": [
                  "to",
                  "from",
                  "subject"
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Persisted and accepted for processing. Not delivery confirmation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "state": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 30
                    }
                  },
                  "required": [
                    "id",
                    "state"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Application": {
        "type": "http",
        "scheme": "bearer"
      },
      "Admin": {
        "type": "http",
        "scheme": "bearer"
      },
      "Internal": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error",
          "requestId"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      }
    }
  }
}